DRAFT: ffiJvm
