+++ app/gui2/shared/util/detect.ts
@@ -5,4 +5,7 @@
 export const isNode =
   typeof global !== 'undefined' && (global as any)[Symbol.toStringTag] === 'global'
 
+// Java environment is set up to have a `jvm: 'graalvm'` property.
+export const isJvm = typeof global !== 'undefined' && (global as any)['jvm'] === 'graalvm'
+
 export const isDevMode = process.env.NODE_ENV === 'development'
