+++ lib/java/js-websocket/src/main/java/org/enso/polyfill/Main.java
@@ -2,8 +2,6 @@ package org.enso.polyfill;
 
 import java.io.File;
 import java.io.IOException;
-import java.io.InputStream;
-import java.nio.file.Path;
 import java.util.concurrent.CompletableFuture;
 import java.util.concurrent.Executors;
