+++ lib/java/js-websocket/src/main/resources/org/enso/polyfill/websocket/websocket-polyfill.js
@@ -279,4 +279,5 @@
 
     globalThis.WebSocketServer = WebSocketServer;
 
+    globalThis.jvm = 'graalvm';
 })
